Major Edits to hex-edit on MacOS: 1. A single click and you can edit the contents of the cell/memory at an address 2. You can edit an address in either hexadecimal, binary, or octal depending on the Radio Button. 3. The base can be represented with u8 integer. So I changed that for every hex-edit no matter the platform.

+++ hiro/windows/widget/hex-edit.cpp
@@ -32,7 +32,7 @@ auto pHexEdit::setAddress(u32 address) -> void {
   update();
 }
 
-auto pHexEdit::setBase(u16 base) -> void {
+auto pHexEdit::setBase(u8 base) -> void {
   update();
 }
